backend/backend/requirements.txt
# Django and REST Framework
Django==5.2.4
djangorestframework==3.16.0

# JWT Authentication
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0

# Additional utilities
django-filter==25.1

# Development dependencies
requests==2.32.3
